requirements.txt
appdirs==1.4.3
attrs==18.2.0
cached-property==1.5.1
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
configparser==3.5.0
defusedxml==0.5.0
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
googleads==15.0.0
httplib2==0.11.3
idna==2.7
isodate==0.6.0
lxml==4.2.5
numpy==1.15.4
oauth2client==4.1.3
oauthlib==2.1.0
public==2.0.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.11.0
suds-jurko==0.6
timedelta==0.0.1
uritemplate==3.0.0
urllib3==1.24.1
xmltodict==0.11.0
zeep==3.1.0
